lec12/hw/requirements.txt
multiprocessing
pillow
requests